Fix codeclimat
